+++ vec2.h
@@ -1,32 +1,34 @@
 #pragma once
 
+#include <types.h>
+
 struct vec2
 {
-        float x, y;
+        uint16 x, y;
 
         vec2();
-        vec2(float scalar);
-        vec2(float x, float y);
+        vec2(uint16 scalar);
+        vec2(uint16 x, uint16 y);
 
         vec2& Add(const vec2& other);
         vec2& Subtract(const vec2& other);
         vec2& Multiply(const vec2& other);
         vec2& Divide(const vec2& other);
 
-        vec2& Add(float value);
-        vec2& Subtract(float value);
-        vec2& Multiply(float value);
-        vec2& Divide(float value);
+        vec2& Add(uint16 value);
+        vec2& Subtract(uint16 value);
+        vec2& Multiply(uint16 value);
+        vec2& Divide(uint16 value);
 
         friend vec2 operator+(vec2 left, const vec2 &right);
         friend vec2 operator-(vec2 left, const vec2 &right);
         friend vec2 operator*(vec2 left, const vec2 &right);
         friend vec2 operator/(vec2 left, const vec2 &right);
 
-        friend vec2 operator+(vec2 left, float value);
-        friend vec2 operator-(vec2 left, float value);
-        friend vec2 operator*(vec2 left, float value);
-        friend vec2 operator/(vec2 left, float value);
+        friend vec2 operator+(vec2 left, uint16 value);
+        friend vec2 operator-(vec2 left, uint16 value);
+        friend vec2 operator*(vec2 left, uint16 value);
+        friend vec2 operator/(vec2 left, uint16 value);
 
         bool operator==(const vec2 &other) const;
         bool operator!=(const vec2 &other) const;
@@ -36,10 +38,10 @@ struct vec2
         vec2& operator*=(const vec2 &other);
         vec2& operator/=(const vec2 &other);
 
-        vec2& operator+=(float value);
-        vec2& operator-=(float value);
-        vec2& operator*=(float value);
-        vec2& operator/=(float value);
+        vec2& operator+=(uint16 value);
+        vec2& operator-=(uint16 value);
+        vec2& operator*=(uint16 value);
+        vec2& operator/=(uint16 value);
 
         bool operator<(const vec2 &other) const;
         bool operator<=(const vec2 &other) const;
@@ -49,5 +51,5 @@ struct vec2
         float Magnitude() const;
         vec2 Normalise() const;
         float Distance(const vec2& other) const;
-        float Dot(const vec2& other) const;
+        uint16 Dot(const vec2& other) const;
 };
